New result structure for read-tree.
For read-tree using a group for node data and for file nodes just being the group seems a reasonable output structure. Each node data gives the source argument to read that the item came from  (ii.e. the parent).  It seems fairly simple to rewrite the tree using visit-tree to other desired tree types as required, include for example adding other node atrtributes.
Not sure that the sequence idea this is built on works very well for variants of building/filtering. One has to rewrite the sequencing function...